requirements.txt
asn1crypto==0.23.0
bcrypt==3.1.4
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.1.3
Django==1.11.7
djangorestframework==3.7.3
Fabric3==1.13.1.post1
idna==2.6
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
paramiko==2.3.1
pyasn1==0.3.7
pycparser==2.18
PyNaCl==1.2.0
pytz==2017.3
requests==2.18.4
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
